requirement.txt
kafka-python
dotenv
pyspark
tweepy
findspark
